requirements.txt
absl-py==2.0.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
easydict==1.11
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.12.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
sympy==1.12
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
